requirements.txt
autopep8==1.4.4
Django==2.2.13
djangorestframework==3.10.3
djangorestframework-jwt==1.8.0
pep8==1.7.1
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2018.5
sqlparse==0.2.4
